requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
Pillow==10.3.0
aiofiles==23.2.1
